added a layer between app and home in order to start Auth and hash routing
